controllers moved to dir
existing controllers moved to dir
created blank controllers for accounts and auth
